Fix repo directory in installation instructions
The `cocrawler` directory is created with the `git clone` if no new directory name is specified. Alternatively `git clone https://github.com/cocrawler/cocrawler.git cocrawl` could be used as the first line the match up with the subsequent instructions but the change in this PR is a bit simpler.